Add files via upload
This assignment required us to use Huffman Encoding to encode strings into binary. Huffman encoding uses the frequency of the characters to determine the optimal encoding of each character in a string. This occurs by creating 'super characters' which represent the characters below it and whose frequency is the sum of the frequency of the characters included in it. First, add all of the characters and frequencies to a dictionary, then use that dictionary to create Nodes (I implemented a simple binary node class) with the corresponding characters and frequencies. Sort the list based upon the frequencies. Combine the 2 nodes with the lowest frequency to make a super node that contains the child nodes in the left and right instance variables (lower frequency goes into left, higher frequency goes into right). Replace the combined nodes with the super node they created and repeat this process until there is only a single node whose character contains all of the characters in the input string. To convert a character from english to binary, traverse the tree made by the single node. If the left child contains the desired character, traverse the left child. If not, traverse the right. Continue traversing until the character of the current node only contains the desired character. For each left traversal, append a 0 to an output string. for each right traversal, append a 1. Return the output string. Repeat this process for all of the characters in the input string. For the sake of bragging, this is the first (and likely last!) time that my code has functioned perfectly upon running it the first time.